Fix deadlock in shutdown method
A deadlock occurred when the BlockingCollection blocked because it
needed another item in order to register the cancellation token's state.
When the cancellation token is cancelled, the BlockingCollection is
already blocked (waiting for another item) but never gets another item
because the OnMessage delegate has been unregistered. This holds onto a
reference to eestream, preventing it from being gc'd and a memory leak
occurred.