Create fixCorruptedLevels.sh
This will remove any corrupted spider level files inside of your currently cd'd directory. It will also remove all open world files, whether they are corrupted or not. Corrupted levels can occur if the program is terminated unexpectedly or does not shut down properly (it happens a lot.) Usually the last line of the file is partially written, so this script will remove that line. If the entire file is corrupted, it will remove the entire file and directory.

It will emit the errors it finds in the files, and will silently correct them. Modification dates are not preserved on corrupted files after running this.